feat: Prioritize exact matches compared to fuzzy
When a  `query` is given, attach will now prioritize exact matches if
there is one exact match. If there is more then one exact match the
fuzzy selector will be used.

This helps when you know the exact thing you want and type some of it
but there is a similar one that just matches when fuzzy.