refactor: move from git2 to gix library for git ops
This change is primarily made to remove the requirement on openssl which
git2 requires. This will allow cross compiling with cross much easier as
cross specifically removed support for openssl.

There were some slight performance improvements with this change as
well. These tests were crude and most operations in tuxmux are io bound
walking the workspace paths but there was a measurable improvement of
around 1ms.

(note the overall time is not representative but the difference is)

hyperfine "./tm-new" "./tm-old"
Benchmark 1: ./tm-new
  Time (mean ± σ):      44.9 ms ±   1.2 ms    [User: 56.3 ms, System: 159.0 ms]
  Range (min … max):    42.0 ms …  48.2 ms    50 runs

Benchmark 2: ./tm-old
  Time (mean ± σ):      69.0 ms ±   3.2 ms    [User: 105.5 ms, System: 259.5 ms]
  Range (min … max):    64.8 ms …  83.4 ms    37 runs

Summary
  './tm-new' ran
    1.54 ± 0.08 times faster than './tm-old'